requirements.txt
requests==2.7.0
GitPython==2.0.6
